Improve usability for compilation errors
We hook the game's package script loading code and pass a string with
the TGI of the resource as the script's file name.
This allows that information to be include in the Lua error messages
that are written to the log file, instead of an empty string.

Note that this feature will only be available on game version 641 due
to the required assembly patching.