Replace the Maxis Lua print function
This allows print to work during script compilation, and allows the
user to opt-in to displaying the Maxis Lua script debug output.

By default, the Maxis Lua script debug output is ignored to avoid
cluttering up the log.
The user can enable or disable it by using the ShowMaxisLuaDebugText
and HideMaxisLuaDebugText cheats in the region view.

Replacing the Maxis Lua print function requires game version 641, for
earlier versions a fallback print function will be added to the advisor
system.
The fall back function will not work during script compilation, but it
should work inside other functions.